pyproject.toml
[tool.poetry]
name = "litellm-proxy-api-client"
version = "1.40.3"
description = "A client library for accessing LiteLLM API"
authors = []
readme = "README.md"
packages = [
    {include = "litellm_proxy_api_client"},
]
include = ["CHANGELOG.md", "litellm_proxy_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
